refactor(ts-sdk): extract Node init helpers + drop stale JSDoc
Resolves my own code review on this PR (3 issues, all flagged at
confidence >= 80):

1) wasm.ts had grown to 510 lines, exceeding the .claude/rules
   code-quality.md per-file 500 NLOC ceiling (Codacy enforces it).
   Move the Node-only helpers (isNodeRuntime + loadWasmBytesNode)
   to a sibling module 'wasm-node-loader.ts'. wasm.ts is now 439
   lines, the new file is 83 lines, both well under 500.

2) The previous JSDoc on loadWasmBytesNode claimed the WASM filename
   was discovered 'by reading @wiscale/velesdb-wasm/package.json#files'.
   The implementation explicitly does the opposite — it reads the
   directory with fs.readdir and skips package.json#files entirely
   (commit 16e30a43). The new doc on wasm-node-loader.ts states the
   actual readdir-based discovery and explains why we deliberately
   avoid the manifest.

3) The same JSDoc claimed createRequire(import.meta.url) 'works under
   both module systems without per-format branching'. The code below
   it has explicit __filename vs import.meta.url branching because
   tsup's CJS output leaves import.meta.url undefined (commit cf37bfae).
   The new doc on wasm-node-loader.ts now describes the per-format
   branching as the actual implementation strategy.

Verification:
- wasm.ts: 510 -> 439 lines.
- 29/29 wasm-backend tests pass.
- ESM + CJS smoke tests in node:20-slim both print
  'first match: id=1 score=1.0000' end-to-end.
